Fix #58 Wrong Patient on Results Reports
https://github.com/bikalabs/bika.health/issues/58

When publishing more than one AR at the same time using the Publish
or Re-publish button from AR lists, the patient from the first selected
AR is set to all results reports.

Added a method in AnalysisRequestsPublish to retrieve the specific
patient for the AR that is currently being rendered by the template.
Also, modifications in the template to fit with this method for
patient retrieval.