Added History Results Tab inside Patient View
In History Results Tab View, the results for the patient are shown
in a table categorized by sample type, analyses and date. The purpose is
to allow the user to see variations amongst different results for the
same patient and analysis service all over the time.
Only results from verified or published analyses are shown.