Show Date Reported in reports
The Date Reported is properly shown when viewing a published report.
However, when previewing it before publishing the field Date Reported
has a blank value before it has not been et yet. If this commit is
applied it will show the current date when the Date Reported isn't set
yet.