Filter doctors by current's user client
If the current user is a Client contact, only the doctors assigned
to the same client and those that do not have any client assigned are
displayed. If user is from the lab, no filtering is applied.

If the current context of the listing is Client, only doctors assigned
to same client are displayed.