Do not allow to change client in Doctor if ARs or Batches
If the current logged in user is from lab personnel, load the
clients as usual, but only if the Doctor does not have any
Analysis Request or Batch assigned. Otherwise, don't allow to
change the Client.

If the user is a client and the Doctor has a client already
set, do not allow to change the primary client.