Improve performance of upgrade step
The patients are moved inside Client objects when running the
upgrade step. Using the classic approach provided by Plone is
slower because performs some permissions checks, serializes the
object to be moved and also wakes-up objects several times
unnecessarily.

With this change, the copy&paste of objects from Patients folder
to Clients is slightly faster, cause some of the actions explained
above are omitted.