Cannot override Patients listing behavior with subscriber adapters
`PatientsView` is used as the "root" listing view for Patients listings. Since
it overrides the function `before_render`, some of the changes made in
subscriber adapters from another add-ons are omitted. For instance, trying to
modify `self.context_actions` via subscriber adapters won't have any effect
because after the call, the value is overrided in PatientsView.

This Pull Request moves the logic `before_render` from the base class
`PatientsView` to the function `update`.