Allow to introduce the Age instead of Date of Birth for Patient
This commit makes the field Age from Patient Add/Edit form to be
visible and editable when the option "Estimated Date of Birth" is
checked, while "Birth Date" field remains hidden. User can then
introduce the age in years, months and days and the date of birth
is then calculated automatically (not accurate).

When the option "Estimated Date of Birth" is unchecked, the age
fields become hidden, while the "Birth Date" becomes visible for
user input.